server/requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
eventlet==0.33.3
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
uuid==1.30
Werkzeug==2.3.4
zope.event==4.6
zope.interface==6.0
